requirements.txt
Django==1.9.4
django-model-utils==2.4
djangorestframework==3.3.2
Pillow==3.1.1
psycopg2==2.6.1
wheel==0.24.0
